requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.7.0
pydantic-settings==2.3.0
python-dotenv==1.0.1
groq==0.9.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
PyMuPDF==1.24.5
pytesseract==0.3.10
Pillow==10.3.0
reportlab==4.2.0
requests==2.32.3
bcrypt==4.1.3
authlib==1.3.1
httpx==0.27.0
